Use TransparentUpgradeableProxy for factory clones
This makes contracts upgradeable along their own axis, independently. To
cause new clones to use a new version, the factory (and thereby the
underlying token as well), would also need to be upgraded. To upgrade
previously cloned tokens, each individual proxy that has been generated
would need to be upgraded. It's somewhat manual and very opt-in.