Implement revocation
A revoke is a burn. No more token.

This has its own role, which the admin is granted by default. This also
necessitated moving transferrable functionality out of the
beforeTransfer callback so as not to allow minters and revokers to have
a back door to transfers when transfers are disabled.